+++ src/components/IssueFields.vue
@@ -1,7 +1,7 @@
 <template>
-    <div v-if="issueFields" class="row q-col-gutter-md items-start" :class="{ 'column': $q.screen.lt.md }">
+    <div v-if="issueFields" class="row items-start">
         <!-- Original Issue Column -->
-        <div class="col-12 col-md-6 q-pr-sm">
+        <div class="col-12 col-md-6 q-pa-sm">
             <div class="text-subtitle1 q-mb-sm">Original Issue Fields</div>
             <q-list separator bordered padding class="rounded-borders">
                 <template v-for="field in issueDisplayFields" :key="field">
@@ -20,32 +20,34 @@
                     </q-item>
                 </template>
             </q-list>
-            <div class="float-right q-ma-sm">
+            <div class="float-right q-mt-sm q-mb-sm">
                 <q-btn color="primary" label="IMPROVE" @click="generateImprovement(props.issueKey)"
                     :loading="loading" />
             </div>
         </div>
 
         <!-- Improvements Column -->
-        <div class="col-12 col-md-6 q-pl-md">
+        <div class="col-12 col-md-6 q-pa-sm">
             <div class="text-subtitle1 q-mb-sm">AI Generated Improvement Proposals</div>
             <div class="description-text">
-                <q-list v-if="improvementProposal && hasImprovements" separator bordered padding
-                    class="rounded-borders">
+
+                <!-- STATE: GENERATING -->
+                <!-- Field text is displayed if updated; comment is always displayed if available -->
+                <q-list v-if="improvementProposal && improvementFieldsFiltered.length > 0" separator bordered padding class="rounded-borders">
                     <template v-for="(field, key) in improvementProposal" :key="key">
-                        <q-item v-if="field?.updated == 'true' || field?.updated == true" style="cursor: default">
+                        <q-item v-if="shouldDisplayField(key)" style="cursor: default">
                             <q-item-section>
                                 <q-item-label class="text-capitalize text-subtitle2">
                                     {{ field?.label || field }}
                                 </q-item-label>
-                                <q-item-label v-if="field?.text">
+                                <q-item-label v-if="field?.text && isFieldUpdated(field)">
                                     <MarkdownViewer :content="field?.text" />
                                 </q-item-label>
                                 <q-item-label v-if="field?.comment" caption class="text-italic q-mt-xs">
                                     Comment: {{ field?.comment }}
                                 </q-item-label>
                             </q-item-section>
-                            <q-item-section side top>
+                            <q-item-section side top v-if="isFieldUpdated(field)">
                                 <q-chip square size="sm" class="text-caption text-uppercase q-ma-none" color="primary"
                                     :clickable="!field.accepted" :outline="field.accepted"
                                     :label="field.accepted ? 'Accepted' : 'Accept'"
@@ -55,39 +57,58 @@
                         </q-item>
                     </template>
                 </q-list>
-                <!-- Failed improvement -->
-                <q-list v-else-if="improvementFailed" bordered padding class="rounded-borders"
-                    style="border-color: var(--q-negative); background-color: rgb(from var(--q-negative) r g b / 10%)">
-                    <q-item>
-                        <q-item-section avatar top>
-                            <q-icon name="mdi-alert-circle-outline" color="negative" />
-                        </q-item-section>
-                        <q-item-section>
-                            <q-item-label class="text-subtitle2">Failed to generate improvements</q-item-label>
-                            <q-item-label>{{ errorMessage }}</q-item-label>
-                        </q-item-section>
-                    </q-item>
-                </q-list>
-                <!-- No improvements needed -->
-                <q-list v-else-if="!improvementFailed && !loading" bordered padding class="rounded-borders"
-                    style="border-color: var(--q-positive); background: rgb(from var(--q-positive) r g b / 10%)">
-                    <q-item>
-                        <q-item-section avatar top>
-                            <q-icon name="mdi-check-circle-outline" color="positive" />
-                        </q-item-section>
-                        <q-item-section>
-                            <q-item-label>All fields look great! No improvements needed.</q-item-label>
-                        </q-item-section>
-                    </q-item>
-                </q-list>
 
-                <!-- Generating improvement -->
+                <!-- STATE: WAITING -->
+                <!-- for streaming respons to start -->
                 <div v-else-if="loading">
                     Generating improvements...
                 </div>
 
-                <!-- First open -->
-                <div v-else class="q-pa-md">Click the improve button to generate improvement proposals</div>
+                <!-- STATE: IDLE -->
+                <div v-else-if="!loading">
+
+                    <!-- STATE: NOT STARTED -->
+                    <q-list v-if="improvementProposal == null" bordered class="rounded-borders"
+                        style="border-color: var(--q-warning); background-color: rgb(from var(--q-warning) r g b / 10%)">
+                        <q-item>
+                            <q-item-section avatar>
+                                <q-icon name="mdi-information-outline" color="warning" />
+                            </q-item-section>
+                            <q-item-section>
+                                <q-item-label>Click the improve button to generate improvement proposals</q-item-label>
+                            </q-item-section>
+                        </q-item>
+                    </q-list>
+
+                    <!-- STATE: NO IMPROVEMENTS AND NO COMMENTS -->
+                    <q-list v-else-if="improvementProposal != null && improvementFieldsFiltered.length == 0" bordered
+                        class="rounded-borders"
+                        style="border-color: var(--q-warning); background-color: rgb(from var(--q-warning) r g b / 10%)">
+                        <q-item>
+                            <q-item-section avatar>
+                                <q-icon name="mdi-information-outline" color="warning" />
+                            </q-item-section>
+                            <q-item-section>
+                                <q-item-label>No improvement proposals or comments</q-item-label>
+                            </q-item-section>
+                        </q-item>
+                    </q-list>
+
+                    <!-- STATE: FAILED -->
+                    <!-- This happens due to connection issues, e.g. if user has no AI credits left.  -->
+                    <q-list v-else-if="improvementFailed" bordered padding class="rounded-borders"
+                        style="border-color: var(--q-negative); background-color: rgb(from var(--q-negative) r g b / 10%)">
+                        <q-item>
+                            <q-item-section avatar>
+                                <q-icon name="mdi-alert-circle-outline" color="negative" />
+                            </q-item-section>
+                            <q-item-section>
+                                <q-item-label class="text-subtitle2">Failed to generate improvements</q-item-label>
+                                <q-item-label>{{ errorMessage }}</q-item-label>
+                            </q-item-section>
+                        </q-item>
+                    </q-list>
+                </div>
             </div>
         </div>
     </div>
@@ -95,7 +116,7 @@
 
 <script setup>
 import { ref, watch, computed } from 'vue';
-import { colors, getCssVar, useQuasar } from 'quasar';
+import { useQuasar } from 'quasar';
 import { useJiraClient } from '../composables/JiraClient.js';
 import { useOpenAIClient } from '../composables/OpenAIClient.js';
 import MarkdownViewer from './MarkdownViewer.vue';
@@ -123,8 +144,6 @@ const { jiraClient } = useJiraClient();
 const { openAIClient } = useOpenAIClient();
 const templateStore = useTemplateStore();
 
-const { textToRgb } = colors
-
 const loading = ref(false);
 const issueFields = ref(null);
 const improvementProposal = ref(null);
@@ -132,17 +151,32 @@ const chunks = ref(0);
 const improvementFailed = ref(false);
 const errorMessage = ref('');
 
+// Fields from the issue to display in the UI
 const issueDisplayFields = ['summary', 'description'];
-const improvementDisplayFields = computed(() => Object.keys(improvementProposal.value || {}));
 
-const shouldDisplayField = (field) => {
-    const fieldData = improvementProposal.value[field];
-    return fieldData?.updated && (field !== 'acceptanceCriteria' || fieldData.text?.length > 0);
+// Fields in the template for the current issue type
+const templateFields = computed(() => {
+    const currentTemplate = templateStore.templates.find(t => t.name === getIssueField('issuetype.name'));
+    return currentTemplate?.fields?.map(field => field.name) || []
+});
+
+// Fields in the improvement proposal
+const improvementFields = computed(() => Object.keys(improvementProposal.value || {}));
+
+// Fields that exist in both the improvementFields and in templateFields
+const improvementFieldsFiltered = computed(() => {
+    return improvementFields.value.filter(field => templateFields.value.includes(field));
+});
+
+// Checks if field with fieldKey from the improvement proposal should be displayed in the UI
+const shouldDisplayField = (fieldKey) => {
+    return improvementFieldsFiltered.value.includes(fieldKey);
 };
 
-const hasImprovements = computed(() => {
-    return improvementDisplayFields.value.some(field => shouldDisplayField(field));
-});
+// Checks if the field with fieldKey from the improvement proposal has been updated
+const isFieldUpdated = (field) => {
+    return field?.updated == 'true' || field?.updated == true;
+};
 
 // Get the value of a field from the issue fields object
 const getIssueField = (field, defaultValue = null) => {
@@ -186,6 +220,7 @@ const generateImprovement = async (issueKey) => {
     loading.value = true;
     chunks.value = 0;
     improvementFailed.value = false;
+    improvementProposal.value = null;
     errorMessage.value = '';
     const issueType = getIssueField('issuetype.name');